references/BearLLM/requirements.txt
torch==2.6.0
scipy>=1.9.0
numpy==1.26.0
python-dotenv==1.0.1
h5pickle==0.4.2
git+https://github.com/huggingface/transformers
git+https://github.com/huggingface/peft